+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `CookiesBanner` now have `CookiesBanner.Button` for the controls. ([@alexandre-lelain](https://github.com/alexandre-lelain))
 * `Footer` now have `Footer.Column` & `Footer.Item` inner components. ([@alexandre-lelain](https://github.com/alexandre-lelain))
 * Renamed `bannerContent` prop of `Footer` to `bottomBanner`. ([@alexandre-lelain](https://github.com/alexandre-lelain))
+* Updated all stories with latest changes. ([@alexandre-lelain](https://github.com/alexandre-lelain))
 
 # 1.0.0