+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.1.0
+
+## Breaking changes
+
+* The Compound Pattern was applied to all components. . ([@alexandre-lelain](https://github.com/alexandre-lelain))
+* `Card` now have `Card.Button` for the controls. . ([@alexandre-lelain](https://github.com/alexandre-lelain))
+* Renamed *onCardClick* prop of `Card` component to *onClick*. ([@alexandre-lelain](https://github.com/alexandre-lelain))
+* `CookiesBanner` now have `CookiesBanner.Button` for the controls. . ([@alexandre-lelain](https://github.com/alexandre-lelain))
+
 # 1.0.0
 
 * First major release of the project. ([@alexandre-lelain](https://github.com/alexandre-lelain))